+++ core/src/poseidon2/mod.rs
@@ -1,5 +1,5 @@
-pub mod baby_bear_gpu;
-pub mod bn254_gpu;
+pub mod baby_bear;
+pub mod bn254;
 pub mod constants;
 
 pub mod tests {
@@ -7,10 +7,10 @@ pub mod tests {
     pub mod baby_bear_tests {
         use crate::device::buffer::DeviceBuffer;
         use crate::device::memory::{ToDevice, ToHost};
-        use crate::poseidon2::baby_bear_gpu::poseidon2_baby_bear_16_kernels::{
+        use crate::poseidon2::baby_bear::poseidon2_baby_bear_16_kernels::{
             DIGEST_WIDTH, D_U64, RATE, ROUNDS_F, ROUNDS_P, WIDTH,
         };
-        use crate::poseidon2::baby_bear_gpu::DeviceHasherBabyBear;
+        use crate::poseidon2::baby_bear::DeviceHasherBabyBear;
         use crate::poseidon2::constants::RC_16_30;
         use p3_baby_bear::{BabyBear, DiffusionMatrixBabyBear};
         use p3_field::{AbstractField, PrimeField32};
@@ -282,10 +282,10 @@ pub mod tests {
     pub mod bn254_tests {
         use crate::device::buffer::DeviceBuffer;
         use crate::device::memory::{ToDevice, ToHost};
-        use crate::poseidon2::bn254_gpu::poseidon2_bn254_3_kernels::{
+        use crate::poseidon2::bn254::poseidon2_bn254_3_kernels::{
             DIGEST_WIDTH, D_U64, RATE, ROUNDS_F, ROUNDS_P, WIDTH,
         };
-        use crate::poseidon2::bn254_gpu::{poseidon2_bn254_3_constants, DeviceHasherBn254};
+        use crate::poseidon2::bn254::{poseidon2_bn254_3_constants, DeviceHasherBn254};
         use p3_bn254_fr::{Bn254Fr, DiffusionMatrixBN254};
         use p3_field::AbstractField;
         use p3_poseidon2::{Poseidon2, Poseidon2ExternalMatrixGeneral};
